requirements.txt
streamlit
pillow
plotly
pandas
sqlite3
google-generativeai
tabulate
ultralytics